feat: impl proxy storage
