+++ crates/rooch-cosmwasm-vm/src/lib.rs
@@ -1,5 +1,15 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod backend;
-pub mod proxy_storage;
+mod move_backend_api;
+mod move_backend_querier;
+mod move_storage;
+mod proxy_storage;
+mod backend;
+
+// Re-export main types from each module
+pub use move_backend_api::MoveBackendApi;
+pub use move_backend_querier::MoveBackendQuerier;
+pub use move_storage::MoveStorage;
+pub use proxy_storage::ProxyStorage;
+pub use backend::build_move_proxy_backend;
